LAMBDO/transformers/requirements.txt
boto3==1.40.11
botocore==1.40.11
jmespath==1.0.1
polars==1.32.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
s3transfer==0.13.1
six==1.17.0
urllib3==2.5.0
uv==0.8.11
